fix: 存在しないコマンドが入力された時の挙動を修正
